refactor: change pages extension, stack navigator refactored
